+++ CHANGELOG.md
@@ -2,3 +2,9 @@
 
 ## [0.1.0] - 17-NOV-2025
 - MVP of AI powered AI concept exploration.
+
+## [0.2.0] - 20-NOV-2025
+- AI concept description update.
+- SSL verification temporary issue fix.
+- K-means clustering update
+- README file improvements.